+++ example/soft_converter_example.dart
@@ -1,8 +1,18 @@
-import 'dart:developer';
+// ignore_for_file: avoid_print
 
 import 'package:soft_converter/soft_converter.dart';
 
-void main() {
-  final awesome = Awesome();
-  log('awesome: ${awesome.isAwesome}');
+Future<void> main() async {
+  final imageConverter = SoftImageConverter();
+
+  try {
+    final file = await imageConverter.toWEBP(
+      input: 'assets/space.jpg',
+      output: 'assets/space.webp',
+    );
+
+    print('Image file: ${file.path}');
+  } catch (e) {
+    rethrow;
+  }
 }