+++ test/soft_converter_test.dart
@@ -1,16 +1,9 @@
-import 'package:soft_converter/soft_converter.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('A group of tests', () {
-    final awesome = Awesome();
+    setUp(() {});
 
-    setUp(() {
-      // Additional setup goes here.
-    });
-
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
-    });
+    test('First Test', () {});
   });
 }